CB-20411 move flow progress methods to a new service
