+++ core/src/main/java/com/sequenceiq/cloudbreak/controller/v4/ProgressV4Controller.java
@@ -11,14 +11,14 @@ import org.springframework.stereotype.Controller;
 import com.sequenceiq.authorization.annotation.CheckPermissionByResourceCrn;
 import com.sequenceiq.authorization.annotation.ResourceCrn;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.progress.ProgressV4Endpoint;
-import com.sequenceiq.cloudbreak.service.progress.ProgressService;
 import com.sequenceiq.flow.api.model.FlowProgressResponse;
+import com.sequenceiq.flow.service.FlowProgressService;
 
 @Controller
 public class ProgressV4Controller implements ProgressV4Endpoint {
 
     @Inject
-    private ProgressService progressService;
+    private FlowProgressService progressService;
 
     @Override
     @CheckPermissionByResourceCrn(action = DESCRIBE_DATALAKE)
