+++ redbeams/src/main/java/com/sequenceiq/redbeams/controller/v4/progress/ProgressV4Controller.java
@@ -12,15 +12,15 @@ import org.springframework.stereotype.Controller;
 import com.sequenceiq.authorization.annotation.CheckPermissionByResourceCrn;
 import com.sequenceiq.authorization.annotation.ResourceCrn;
 import com.sequenceiq.flow.api.model.FlowProgressResponse;
+import com.sequenceiq.flow.service.FlowProgressService;
 import com.sequenceiq.redbeams.api.endpoint.v4.progress.ProgressV4Endpoint;
-import com.sequenceiq.redbeams.service.progress.ProgressService;
 
 @Controller
 @Transactional(Transactional.TxType.NEVER)
 public class ProgressV4Controller implements ProgressV4Endpoint {
 
     @Inject
-    private ProgressService progressService;
+    private FlowProgressService progressService;
 
     @Override
     @CheckPermissionByResourceCrn(action = DESCRIBE_DATABASE_SERVER)
